+++ src/renderer/src/utils/export/utils.ts
@@ -1,0 +1,171 @@
+import { loggerService } from '@logger'
+import i18n from '@renderer/i18n'
+import { getProviderLabel } from '@renderer/i18n/label'
+import { getMessageTitle } from '@renderer/services/MessagesService'
+import store from '@renderer/store'
+import { setExportState } from '@renderer/store/runtime'
+import type { Topic } from '@renderer/types'
+import type { Message } from '@renderer/types/newMessage'
+import { removeSpecialCharactersForFileName } from '@renderer/utils/file'
+import { convertMathFormula, markdownToPlainText } from '@renderer/utils/markdown'
+import { getCitationContent, getMainTextContent, getThinkingContent } from '@renderer/utils/messageUtils/find'
+import dayjs from 'dayjs'
+import DOMPurify from 'dompurify'
+
+const logger = loggerService.withContext('Utils:export')
+const getExportState = () => store.getState().runtime.export.isExporting
+const setExportingState = (isExporting: boolean) => {
+  store.dispatch(setExportState({ isExporting }))
+}
+
+  // 先处理换行符转换为 <br>
+  const contentWithBr = content.replace(/\n/g, '<br>')
+
+  // 使用 DOMPurify 清理内容，保留常用的安全标签和属性
+  return DOMPurify.sanitize(contentWithBr, {
+    ALLOWED_TAGS: [
+      // 换行和基础结构
+      'br',
+      'p',
+      'div',
+      'span',
+      // 文本格式化
+      'strong',
+      'b',
+      'em',
+      'i',
+      'u',
+      's',
+      'del',
+      'mark',
+      'small',
+      // 上标下标（数学公式、引用等）
+      'sup',
+      'sub',
+      // 标题
+      'h1',
+      'h2',
+      'h3',
+      'h4',
+      'h5',
+      'h6',
+      // 引用
+      'blockquote',
+      // 列表
+      'ul',
+      'ol',
+      'li',
+      // 代码相关
+      'code',
+      'pre',
+      'kbd',
+      'var',
+      'samp',
+      // 表格（AI输出中可能包含表格）
+      'table',
+      'thead',
+      'tbody',
+      'tfoot',
+      'tr',
+      'td',
+      'th',
+      // 分隔线
+      'hr'
+    ],
+    ALLOWED_ATTR: [
+      // 安全的通用属性
+      'class',
+      'title',
+      'lang',
+      'dir',
+      // code 标签的语言属性
+      'data-language',
+      // 表格属性
+      'colspan',
+      'rowspan',
+      // 列表属性
+      'start',
+      'type'
+    ],
+    KEEP_CONTENT: true, // 保留被移除标签的文本内容
+    RETURN_DOM: false,
+    SANITIZE_DOM: true,
+    // 允许的协议（预留，虽然目前没有允许链接标签）
+    ALLOWED_URI_REGEXP: /^(?:(?:(?:f|ht)tps?):|[^a-z]|[a-z+.-]+(?:[^a-z+.-:]|$))/i
+  })
+}
+
+/**
+ * 获取话题的消息列表，使用TopicManager确保消息被正确加载
+ * 这样可以避免从未打开过的话题导出为空的问题
+ * @param topicId 话题ID
+ * @returns 话题消息列表
+ */
+async function fetchTopicMessages(topicId: string): Promise<Message[]> {
+  const { TopicManager } = await import('@renderer/hooks/useTopic')
+  return await TopicManager.getTopicMessages(topicId)
+}
+
+/**
+ * 从消息内容中提取标题，限制长度并处理换行和标点符号。用于导出功能。
+ * @param {string} str 输入字符串
+ * @param {number} [length=80] 标题最大长度，默认为 80
+ * @returns {string} 提取的标题
+ */
+export function getTitleFromString(str: string, length: number = 80): string {
+  let title = str.trimStart().split('\n')[0]
+
+  if (title.includes('。')) {
+    title = title.split('。')[0]
+  } else if (title.includes('，')) {
+    title = title.split('，')[0]
+  } else if (title.includes('.')) {
+    title = title.split('.')[0]
+  } else if (title.includes(',')) {
+    title = title.split(',')[0]
+  }
+
+  if (title.length > length) {
+    title = title.slice(0, length)
+  }
+
+  if (!title) {
+    title = str.slice(0, length)
+  }
+
+  return title
+}
+
+const getRoleText = (role: string, modelName?: string, providerId?: string): string => {
+  const { showModelNameInMarkdown, showModelProviderInMarkdown } = store.getState().settings
+
+  if (role === 'user') {
+    return '🧑‍💻 User'
+  } else if (role === 'system') {
+    return '🤖 System'
+  } else {
+    let assistantText = '🤖 '
+    if (showModelNameInMarkdown && modelName) {
+      assistantText += `${modelName}`
+      if (showModelProviderInMarkdown && providerId) {
+        const providerDisplayName = getProviderLabel(providerId) ?? providerId
+        assistantText += ` | ${providerDisplayName}`
+        return assistantText
+      }
+      return assistantText
+    } else if (showModelProviderInMarkdown && providerId) {
+      const providerDisplayName = getProviderLabel(providerId) ?? providerId
+      assistantText += `Assistant | ${providerDisplayName}`
+      return assistantText
+    }
+    return assistantText + 'Assistant'
+  }
+}
+
+/**
+ * 处理文本中的引用标记
+ * @param content 原始文本内容
+ * @param mode 处理模式：'remove' 移除引用，'normalize' 标准化为Markdown格式
+ * @returns 处理后的文本
+ */
+export const processCitations = (content: string, mode: 'remove' | 'normalize' = 'remove'): string => {
